+++ src/flows/ContractorOnboarding/components/PricingPlan.tsx
@@ -42,9 +42,9 @@ export function PricingPlanStep({
 
   const handleSubmit = async (payload: $TSFixMe) => {
     try {
-      await onSubmit?.(
-        contractorOnboardingBag?.parseFormValues(payload) as $TSFixMe, // TODO: add type
-      );
+      const parsedValues =
+        await contractorOnboardingBag?.parseFormValues(payload);
+      await onSubmit?.(parsedValues as $TSFixMe);
       const response = await contractorOnboardingBag?.onSubmit(payload);
       if (response?.data) {
         await onSuccess?.(response?.data as $TSFixMe); // TODO: add type