+++ src/flows/Termination/PaidTimeOffForm.tsx
@@ -15,9 +15,8 @@ type PaidTimeOffFormProps = {
 export function PaidTimeOffForm({ onSubmit }: PaidTimeOffFormProps) {
   const { terminationBag } = useTerminationContext();
   const handleSubmit = async (values: TerminationFormValues) => {
-    await onSubmit?.(
-      terminationBag?.parseFormValues(values) as PaidTimeOffFormValues,
-    );
+    const parsedValues = await terminationBag?.parseFormValues(values);
+    await onSubmit?.(parsedValues as PaidTimeOffFormValues);
     terminationBag?.next();
   };